rename fs/folder to fs/browser suggested by jzila
