+++ shared/fs/banner/reset-banner/container.js
@@ -9,7 +9,7 @@ import * as RPCTypes from '../../../constants/types/rpc-gen'
 import {namedConnect, isMobile} from '../../../util/container'
 import {folderNameWithoutUsers} from '../../../util/kbfs'
 import Banner, {getHeight} from '.'
-import * as RowTypes from '../../folder/rows/types'
+import * as RowTypes from '../../browser/rows/types'
 
 type OwnProps = {|path: Types.Path|}
