+++ shared/fs/browser/destination-picker/container.js
