+++ shared/fs/browser/destination-picker/index.js
