+++ shared/fs/browser/offline.js
