+++ shared/fs/browser/rows/editing-container.js
