+++ shared/fs/browser/rows/load-files-when-needed.js
