+++ shared/fs/browser/rows/uploading-container.js
