+++ shared/fs/browser/rows/uploading.js
