+++ shared/fs/index.stories.js
@@ -2,11 +2,11 @@
 import bannerStories from './banner/index.stories'
 import commonStories from './common/index.stories'
 import filepreviewStories from './filepreview/index.stories'
-import folderStories from './folder/index.stories'
+import browserStories from './browser/index.stories'
 import footerStories from './footer/index.stories'
 import navHeaderStories from './nav-header/index.stories'
 import topBarStories from './top-bar/index.stories'
-import rowStories from './folder/rows/index.stories'
+import rowStories from './browser/rows/index.stories'
 import sendAttachmentToChatStories from './send-to-chat/attachment/index.stories'
 import sendLinkToChatStories from './send-to-chat/link/index.stories'
 import simpleScreensStories from './simple-screens/index.stories'
@@ -16,7 +16,7 @@ export default () =>
     bannerStories,
     commonStories,
     filepreviewStories,
-    folderStories,
+    browserStories,
     footerStories,
     navHeaderStories,
     topBarStories,
