+++ shared/fs/top-bar/index.js
@@ -3,7 +3,7 @@ import * as React from 'react'
 import * as Styles from '../../styles'
 import * as Kb from '../../common-adapters'
 import * as Types from '../../constants/types/fs'
-import * as RowTypes from '../folder/rows/types'
+import * as RowTypes from '../browser/rows/types'
 import Loading from './loading'
 import Sort from './sort-container'
 import SyncToggle from './sync-toggle-container'
